feat(api): useLocationQueryStateItem -> useLocationQueryState (default export)
the single item api is the recommended default

BREAKING CHANGE: exports were renamed, please check documentation

fix #28
