+++ src/examples/use-location-state/01-simple/package.json
@@ -8,7 +8,7 @@
     "add": "^2.0.6",
     "node-sass": "^4.11.0",
     "react-scripts": "3.0.0",
-    "typescript": "3.3.3",
+    "typescript": "3.4.4",
     "yarn": "^1.15.2"
   },
   "dependencies": {
