+++ src/packages/query-state-core/package.json
@@ -12,7 +12,7 @@
     "rollup": "^1.0.0",
     "rollup-plugin-typescript2": "^0.21.0",
     "tslib": "^1.9.3",
-    "typescript": "^3.4.4"
+    "typescript": "3.4.4"
   },
   "files": [
     "dist"
