+++ src/packages/use-location-state/package.json
@@ -20,7 +20,7 @@
     "rollup": "^1.0.0",
     "rollup-plugin-typescript2": "^0.21.0",
     "tslib": "^1.9.3",
-    "typescript": "^3.4.4"
+    "typescript": "3.4.4"
   },
   "files": [
     "dist"
